README.md
This project is a simple but powerful Smart Parking Lot Management System built in C by Sajib Kundu. It's a console app that helps manage a parking lot easily. This system is different because it's not just a basic project; it has many real-world features like a reservation system, payment calculation, and daily reports. It's designed with a clean, organized structure, making the code easy to understand. This system is a great example of how to make a complete and useful app in C that is both affordable and easy to use.